fix some forloop range bugs

+++ HW4/calculateSAD.m
@@ -1,10 +1,11 @@
 function [ SAD ] = calculateSAD( img1, img2, N )
 %UNTITLED4 Summary of this function goes here
 %   Detailed explanation goes here
-    SAD = 0;
+    SAD = uint64(0);
+   
     for i = 1:N,
         for j = 1:N,
-            SAD = SAD + abs(img1(i,j) - img2(i,j));
+            SAD = SAD + uint64(abs(img1(i,j) - img2(i,j)));
         end
     end
